+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [1.5.2] - 2025-01-30
+- Fixed issue where loop object would only show autocomplete suggestions when it was inside `{{` and `}}` brackets
+
 ## [1.5.1] - 2025-01-30
 - Added missing autocomplete for directives `@for`, `@for @else`, `@break`, `@breakIf`, `@continue`, `@continueIf`
 - Improve directive autocomplete descriptions